+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file
 ##[3.1]
 ###Added
 - Insecure option to the REST call constructs
+- (Un)escape XML constructs
 ###Fixed
 - Issue where focus would be lost on the project view [CTC-1254](https://xillio.atlassian.net/browse/CTC-1254)